Re-arrange local-tx-submission code
