+++ pallas-network/src/miniprotocols/localtxsubmission/cardano_node_errors.rs
@@ -1,161 +1,20 @@
+//! This modules contains Rust-equivalents of Haskell types from `cardano-ledger` representing
+//! errors that are sent from the cardano node in the local-TX-submission miniprotocol.
 use pallas_codec::minicbor::{
     self,
     data::Type,
-    decode::{self, Error, Token},
+    decode::{Error, Token},
     Decode, Decoder, Encode,
 };
 use pallas_primitives::conway::ScriptHash;
 use pallas_utxorpc::TxHash;
 
-use crate::miniprotocols::localtxsubmission::{codec::DecodingResult, Message};
-
-use super::{codec::DecodeCBORSplitPayload, EraTx};
-
-/// Decodes Cardano node errors whose CBOR byte representation could be split over multiple
-/// payloads.
-pub struct NodeErrorDecoder {
-    /// When decoding the error responses of the node, we use a stack to track the location of the
-    /// decoding relative to an outer scope (most often a definite array). We need it because if we
-    /// come across an error that we cannot handle, we must still consume all the CBOR bytes that
-    /// represent this error.
-    pub context_stack: Vec<OuterScope>,
-    /// Response bytes from the cardano node. Note that there are payload limits and so the bytes
-    /// may be truncated.
-    pub response_bytes: Vec<u8>,
-    /// This field is used to determine if there are still CBOR bytes that have yet to be decoded.
-    ///
-    /// It has a value of 0 if decoding has not yet started. Otherwise it takes the value of the
-    /// index in `response_bytes` that is also pointed to by the minicbor decoder after a
-    /// _successful_ decoding of a `TxApplyErrors` instance.
-    pub ix_start_unprocessed_bytes: usize,
-    /// This field is true if the current decoding of a `TXApplyErrors` instance is complete, which
-    /// only happens once the CBOR BREAK token is decoded to terminate the indefinite array which is
-    /// part of the `TxApplyErrors` encoded structure.
-    pub cbor_break_token_seen: bool,
-}
-
-impl NodeErrorDecoder {
-    pub fn new() -> Self {
-        Self {
-            context_stack: vec![],
-            response_bytes: vec![],
-            ix_start_unprocessed_bytes: 0,
-            cbor_break_token_seen: false,
-        }
-    }
-}
-
-impl Default for NodeErrorDecoder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl DecodeCBORSplitPayload for NodeErrorDecoder {
-    type Entity = Message<EraTx, Vec<ApplyTxError>>;
-
-    fn try_decode_with_new_bytes(
-        &mut self,
-        bytes: &[u8],
-    ) -> Result<DecodingResult<Self::Entity>, decode::Error> {
-        if self.has_undecoded_bytes() {
-            self.response_bytes.extend_from_slice(bytes);
-            let bytes = self.response_bytes.clone();
-            let mut decoder = Decoder::new(&bytes);
-            let mut errors = vec![];
-
-            loop {
-                match ApplyTxError::decode(&mut decoder, self) {
-                    Ok(tx_err) => {
-                        errors.push(tx_err);
-                    }
-                    Err(e) => {
-                        if !e.is_end_of_input() {
-                            return Err(e);
-                        } else {
-                            break;
-                        }
-                    }
-                }
-            }
-
-            if self.has_undecoded_bytes() {
-                Ok(DecodingResult::Incomplete(Message::RejectTx(errors)))
-            } else {
-                Ok(DecodingResult::Complete(Message::RejectTx(errors)))
-            }
-        } else {
-            // If it's not an error response then process it right here and return.
-            let mut d = Decoder::new(bytes);
-            let mut probe = d.probe();
-            if probe.array().is_err() {
-                // If we don't have any unprocessed bytes the first element should be an array
-                return Err(decode::Error::message(
-                    "Expecting an array (no unprocessed bytes)",
-                ));
-            }
-            let label = probe.u16()?;
-            match label {
-                0 => {
-                    d.array()?;
-                    d.u16()?;
-                    let tx = d.decode()?;
-                    Ok(DecodingResult::Complete(Message::SubmitTx(tx)))
-                }
-                1 => Ok(DecodingResult::Complete(Message::AcceptTx)),
-                2 => {
-                    self.response_bytes.extend_from_slice(bytes);
-                    let bytes = self.response_bytes.clone();
-                    let mut decoder = Decoder::new(&bytes);
-                    let mut errors = vec![];
-
-                    loop {
-                        match ApplyTxError::decode(&mut decoder, self) {
-                            Ok(tx_err) => {
-                                errors.push(tx_err);
-                            }
-                            Err(e) => {
-                                if !e.is_end_of_input() {
-                                    return Err(e);
-                                } else {
-                                    break;
-                                }
-                            }
-                        }
-                    }
-
-                    if self.has_undecoded_bytes() {
-                        Ok(DecodingResult::Incomplete(Message::RejectTx(errors)))
-                    } else {
-                        Ok(DecodingResult::Complete(Message::RejectTx(errors)))
-                    }
-                }
-                3 => Ok(DecodingResult::Complete(Message::Done)),
-                _ => Err(decode::Error::message("can't decode Message")),
-            }
-        }
-    }
-
-    fn has_undecoded_bytes(&self) -> bool {
-        self.ix_start_unprocessed_bytes + 1 < self.response_bytes.len()
-    }
-}
+use super::codec::NodeErrorDecoder;
 
 /// https://github.com/IntersectMBO/cardano-ledger/blob/8fd7ab6ca9bcf9cdb1fa6f4059f84585a084efa5/eras/shelley/impl/src/Cardano/Ledger/Shelley/API/Mempool.hs#L221
 #[derive(Debug, Clone)]
 pub struct ApplyTxError {
     pub node_errors: Vec<ShelleyLedgerPredFailure>,
-}
-
-impl Encode<()> for ApplyTxError {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        _e: &mut minicbor::Encoder<W>,
-        _ctx: &mut (),
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        // We only ever decode node errors.
-        unreachable!()
-    }
 }
 
 impl Decode<'_, NodeErrorDecoder> for ApplyTxError {
@@ -219,6 +78,17 @@ impl Decode<'_, NodeErrorDecoder> for ApplyTxError {
             }
         }
 
+        unreachable!()
+    }
+}
+
+impl Encode<()> for ApplyTxError {
+    fn encode<W: minicbor::encode::Write>(
+        &self,
+        _e: &mut minicbor::Encoder<W>,
+        _ctx: &mut (),
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        // We only ever decode node errors.
         unreachable!()
     }
 }
