+++ pallas-network/src/miniprotocols/localtxsubmission/mod.rs
@@ -1,4 +1,5 @@
 pub use client::*;
+pub use codec::NodeErrorDecoder;
 pub use protocol::*;
 
 pub mod cardano_node_errors;
